Initial reworking of sensor task for TSL2585
Added code to configure and capture readings from the new sensor.
Updated the diagnostic function to show data from the new sensor.

Renamed a lot of the old functions and types to avoid having to
break/rewrite most of the measurement code just yet.

Right now this is a partial update, and focus will be on the diagnostic
code paths until more higher level logic can be rewritten around the new
sensor.