Initial update to desktop app to detect device type
The plan for the desktop app will be to try and have a single app
support both the UV/VIS device and the original device. At some point,
this version of the desktop app will be merged back into the baseline or
forked off as an independent project.

So far device type is detected from the version string, and the only
device-specific features added are the temperature sensor and UV LED
control.