Add ability to change sensor photodiode configurations
This makes it possible to explicitly set Photopic-only or UV-only
measurement modes.

At the moment running both types of photodiodes at the same time,
through different modulators, is not supported. This could be added, but
it is likely that these two will be used independently from each other
for cleaner measurements.