Add initial use of UV calibration value
This is applied to the raw data coming off the sensor, and is
effectively a linear adjustment per the datasheet.  Without more
information, its unclear how much of a difference this will actually
make.  Also, it might make more sense to apply this to a higher level
value, but we'll see.