Initial CMake support for the main firmware and bootloader
This also includes various IDE-specific files for CLion

One important gotcha that was discovered, is that the order of the
linker arguments actually matters. More specifically, the arguments need
to be passed after the object files to avoid a lot of linker errors that
crop up. This is covered more in this thread:

https://community.arm.com/support-forums/f/compilers-and-libraries-forum/54147/gcc-12-2-libc_nano-exporting-retargetable-locking-functions-and-data