Minor fixes and improvements to capturing sensor readings
Changed the way we handle the FreeRTOS scheduler when reading sensor
data to avoid unnecessary delays from task switching.

Added a workaround to an issue where, after a gain change, data would
continue to be returned before the gain change had fully taken effect.
This was breaking gain calibration, sometimes.

This bug still needs to be fixed for real, but the workaround added to
the gain calibration code is good enough for the time being.