CMake integration improvements
Rework the way CMake is used in the project such that all the external
bits have their own CMake configurations and are linked as intermediary
libraries.

This provides a more structured build process, and lets CMake handle
some of the dependency management.  It is also similar to how the
Printalyzer Timer project is now built.

This will also make it easier to integrate libraries that provide their
own CMake configurations.