Add oscillator calibration to sensor startup
The Rev C hardware now has a 128Hz reference signal routed to the
TSL2585 sensor's VSYNC pin. This change adds the code to use that signal
to perform sensor oscillator calibration on sensor startup.  This
routine is fairly quick, and will run every time the sensor is started
to take a measurement.

The one downside of this change is that the sensor will now fail to
start if the signal is absent, so the related code must be commented out
prior to testing on Rev B hardware. This could be worked around with
better error handling in the future, or just left in if we assume this
feature will remain going forward.