Sensor task implementation should reflect light mutual exclusivity
The control API for the measurement lights was written to enforce the
three different lights being mutually exclusive, but the actual internal
implementation still had them as three separate values.

This change updates all the internal state variables to reflect only one
light setting being possible at a time, and it also enforces an order
where all non-active lights are disabled before the active light is
enabled. This is necessary because enabling multiple light sources can
cause the device to exceed its current limits.

This change should probably be propagated to the lower-level LED driver
runner code, but as that code is never called outside of the sensor
task, it is probably fine to just do this inside the sensor task.