Remove measurement light calibration settings
Ideally all density measurements should be taken with the measurement
light at its maximum power with no PWM. However, due to manufacturing
variations, this ended up violating some assumptions of the gain
selection process on the DPD-100. As such, a calibrated max light value
was put in as a last-minute fix to avoid having to make any rushed
changes to such a critical code path.

With the newer UV/VIS device, gain selection is a completely different
process and this issue does not exist. It is therefore always safe to
just use the maximum brightness for measurement. So there is no point in
leaving in any of the old code that was there to handle this situation.

The space in EEPROM where these settings were stored has been left in as
a reserved section, in case we ever need to put something else there in
the future.