Expand remote reading commands
Added a command to do a full remote measurement cycle, with all the same
gain detection logic as a user-triggered measurement. This is to support
device characterization experiments where we'd rather not hard-code
sensor settings. Only issue is that this returns basic counts, and thus
won't give useful data unless gain calibration has already been
performed.

Added arguments to both the remote read and remote measure commands to
allow for explicitly setting the light brightness. Normally these will
be set to the max value, but there are diagnostic cases where it may be
desired to change them.