Remove crystal reference and reroute VSYNC pin
In testing with Rev C hardware, it was determined that using the VSYNC
pin to calibrate the TSL2585's internal oscillator can actually
introduce more instability in sensor timing than would otherwise exist.
Furthermore, the most likely factor affecting oscillator stability is
temperature, which we're already planning to correct for.

Thus these extra components should be removed from the design.

For the sake of future-proofing, the VSYNC pin is now rounded directly
to the MCU (though a pull-up resistor), in case someday it is decided to
use this as a measurement trigger. It is also likely this pin will never
be used, but it doesn't really hurt to do this.

Fix #2