Remove sensor oscillator calibration code
It has been decided to remove this feature from the hardware, so it
should also be removed from the firmware.

Calibration the sensor's oscillator to an external reference sounds nice
in theory, but there's enough error in the process that it actually
makes the sensor timing less stable if done regularly.

Going forward, if the sensor's VSYNC pin is used for anything at all, it
should only be used to trigger readings. And even that is not too likely
due to overhead.