Minor changes for a quick order of a copper substrate variant
If this ends up being the final version of the transmission LED board,
then it should be merged back into master with notes and cleanup to make
the changes obvious.

Right now its just the bare minimum of changes to support JLCPCB's
copper PCB service, with some aspects being in the manufacturing notes
and not the KiCad files.