Quick fix for string width calculation bug in u8g2
This bug is better described in the upstream issue:
https://github.com/olikraus/u8g2/issues/2588

Basically there can be minor alignment errors in menu text because of
how the position is calculated. This was not seen before, because it
pretty much requires the first character of the text to be narrow and
the text to span the full width of the display.