Pick final PWM frequency and make more calculations dynamic
The LED PWM frequency is a compromise between getting good stable
readings, and maintaining a 1:1 relationship (on a logarithmic scale)
between duty cycle and measured brightness. This commit reflects that
decision and adds some relevant comments.

We should now be able to reliably go down to a 1% duty cycle, or an OD
equivalent of 2.00, while maintaining this relationship.  If we ever
need to go further, a dynamic frequency (and some measurement timing
changes) will be required.