Gain calibration process rewrite
The creative approach to gain calibration that was originally
implemented, did not produce the greatest results when compared to a
more "correct" manual filter-based calibration.

Fixing this involved a complete change to how gain calibration is
performed, with the new approach giving similar results to a
filter-based calibration.

The first step was to change the LED PWM frequency during gain
calibration for better sensor behavior.  This means the timer running
that signal is now configurable, using a high-frequency/low-range
setting for gain calibration and a low-frequency/high-range setting for
normal operation.

The next step was to separate the process of finding the per-gain light
setting into a separate routine, converging on the hopefully ideal
setting for measuring each gain pair.

Finally, for actually measuring each gain pair, the transmission LED is
no longer kept continuously on.  To make behavior more consistant and
repeatable, it is now used in a normal measurment cycle.  An individual
measurement cycle is performed at each necessary gain/light setting,
with some cooldowm time in-between, and the results are appropriately
compared.