Remove slope/linearity calibration features
Extensive testing with various reference materials has shown that the
slope calibration features, which were mostly inherited from the
DPD-100, are probably best avoided going forward.

The nature of any corrections are highly dependent on exactly which
reference materials are used, and are thus not a trustworthy calibration
of general sensor behavior. Furthermore, if corrections are ever
re-added, they likely would need to be measurement-mode-specific for
this reason.

Furthermore, the primary driver for the slope corrections on the DPD-100
was a low-density error curve that was observed in testing.  This was
also observed on the DPD-105 (this device) when measuring film with the
emulsion facing up. However, when flipping the flim to be emulsion facing
down, this error curve vanishes from the data.

So with film measured emulsion-down, linearity is sufficiently good
across the range when testing against a variety of materials.

Additionally, by removing this bit of calibration data, it becomes much
easier for end-users to recalibrate their devices against alternative
reference materials.

(Temperature correction still uses a polynomial slope correction
approach, but it is based on measurments taken by modulating the LED
brightness. When done correctly, those produce significantly cleaner
data than any film measurements.)