Increase size of CDC handler task stack
Lately there have been some spurious device resets when controlling it
through the USB-CDC interface. Further investigating has shown that task
overflows on the "cdc" task sometimes happen, and the watermark was
awfully close to the limit.

To try and solve this, the stack size of that task has been increased.