Simplify the temperature correction process
This gets rid of the overly complex and hard-to-explain approach to
correcting for temperature with multiple independent variables, and
moves towards a much simpler approach.  The new approach uses a simple
polynomial regression on a single indepdendent variable to correct for the
relationship between temperature and reading.

This commit itself updates the USB-CDC configuration protocol, the
settings data structures, and the correction function, to support this
change.