Add support for negative reflection CAL-LO density values
Support for this is being added to settings, density calculations, and
the display site of the calibration menu. There is no way to actually
input these values on-device, so that will have to be done using the
desktop software.

Certain lab reflectance standards (notably PTFE-based ones) may have a
reflectance factor >1 when measured with a 0/45 geometry. This
corresponds to a negative reflection density. As these standards may be
used for upstream calibration of reflection references, some minimal
support needs to be added to the device firmware.